refactor(chat): isolate attachment cards
